+++ CHANGELOG.md
@@ -31,6 +31,9 @@
 * [#754](https://github.com/mozilla/glean.js/pull/754): Change target ECMAScript target from 2015 to 2016 when building for Qt.
 * [#779](https://github.com/mozilla/glean.js/pull/779): Add a number of workarounds for the Qt Javascript engine.
 
+* [#775](https://github.com/mozilla/glean.js/pull/775): Disallow calling test only methods outside of test mode.
+  * NOTE: Test mode is set once the API `Glean.testResetGlean` is called.
+
 # v0.20.0 (2021-09-17)
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v0.19.0...v0.20.0)